Refactor timeout and buffer time types in coinflip program
- Changed the data type for `oracle_buffer_time`, `max_timeout`, `min_timeout`, `timeout`, and related fields from `i64` to `u16` across multiple structs and functions, improving memory efficiency and consistency.
- Updated the `calculate_winner` method to accept a `random_number` instead of a hash value, enhancing clarity in winner determination.
- Removed unused comments and streamlined fee calculation logic in the `set_oracle_hash` instruction.

These changes enhance the maintainability and clarity of the coinflip program by optimizing data types and simplifying logic.

+++ programs/coinflip/src/instructions/initialize_game.rs
@@ -9,7 +9,7 @@ pub fn handler(
     amount: u64,
     max_players: u16,
     min_players: u16,
-    timeout: i64,
+    timeout: u16,
     is_private: bool,
 ) -> Result<()> {
     // Initialize game
@@ -23,7 +23,7 @@ pub fn handler(
     game.players = Vec::with_capacity(max_players as usize);
     game.status = crate::state::GameStatus::Active;
     game.token_mint = ctx.accounts.game_token.token_mint;
-    game.created_at = Clock::get()?.unix_timestamp;
+    game.created_at = Clock::get()?.unix_timestamp as u64;
     game.timeout = timeout;
     game.is_private = is_private;
